+++ src/client/conn.rs
@@ -10,10 +10,11 @@
 use std::fmt;
 use std::marker::PhantomData;
 use std::mem;
+use std::sync::Arc;
 
 use bytes::Bytes;
 use futures::{Async, Future, Poll};
-use futures::future::{self, Either};
+use futures::future::{self, Either, Executor};
 use tokio_io::{AsyncRead, AsyncWrite};
 
 use body::Payload;
@@ -70,9 +71,10 @@ where
 /// After setting options, the builder is used to create a `Handshake` future.
 #[derive(Clone, Debug)]
 pub struct Builder {
-    exec: Exec,
+    pub(super) exec: Exec,
     h1_writev: bool,
     h1_title_case_headers: bool,
+    h1_read_buf_exact_size: Option<usize>,
     http2: bool,
 }
 
@@ -431,13 +433,18 @@ impl Builder {
         Builder {
             exec: Exec::Default,
             h1_writev: true,
+            h1_read_buf_exact_size: None,
             h1_title_case_headers: false,
             http2: false,
         }
     }
 
-    pub(super) fn exec(&mut self, exec: Exec) -> &mut Builder {
-        self.exec = exec;
+    /// Provide an executor to execute background HTTP2 tasks.
+    pub fn executor<E>(&mut self, exec: E) -> &mut Builder
+    where
+        E: Executor<Box<Future<Item=(), Error=()> + Send>> + Send + Sync + 'static,
+    {
+        self.exec = Exec::Executor(Arc::new(exec));
         self
     }
 
@@ -451,6 +458,10 @@ impl Builder {
         self
     }
 
+    pub(super) fn h1_read_buf_exact_size(&mut self, sz: Option<usize>) -> &mut Builder {
+        self.h1_read_buf_exact_size = sz;
+        self
+    }
     /// Sets whether HTTP2 is required.
     ///
     /// Default is false.
@@ -495,6 +506,9 @@ where
             }
             if self.builder.h1_title_case_headers {
                 conn.set_title_case_headers();
+            }
+            if let Some(sz) = self.builder.h1_read_buf_exact_size {
+                conn.set_read_buf_exact_size(sz);
             }
             let cd = proto::h1::dispatch::Client::new(rx);
             let dispatch = proto::h1::Dispatcher::new(cd, conn);