+++ src/server/tcp.rs
@@ -8,7 +8,7 @@ use tokio_reactor::Handle;
 use tokio_tcp::TcpListener;
 use tokio_timer::Delay;
 
-use self::addr_stream::AddrStream;
+pub use self::addr_stream::AddrStream;
 
 /// A stream of connections from binding to an address.
 #[must_use = "streams do nothing unless polled"]
@@ -29,7 +29,7 @@ impl AddrIncoming {
         if let Some(handle) = handle {
             AddrIncoming::from_std(std_listener, handle)
         } else {
-            let handle = Handle::current();
+            let handle = Handle::default();
             AddrIncoming::from_std(std_listener, &handle)
         }
     }
@@ -46,6 +46,11 @@ impl AddrIncoming {
             tcp_nodelay: false,
             timeout: None,
         })
+    }
+
+    /// Creates a new `AddrIncoming` binding to provided socket address.
+    pub fn bind(addr: &SocketAddr) -> ::Result<Self> {
+        AddrIncoming::new(addr, None)
     }
 
     /// Get the local address bound to this listener.
@@ -194,6 +199,7 @@ mod addr_stream {
     use tokio_io::{AsyncRead, AsyncWrite};
 
 
+    /// A transport returned yieled by `AddrIncoming`.
     #[derive(Debug)]
     pub struct AddrStream {
         inner: TcpStream,